Restructuration du contrôle des données
